Assistance/requirements.txt
paho-mqtt
pyaudio
SpeechRecognition